+++ src/main/java/com/jof/springmvc/dao/ReviewDaoImpl.java
@@ -1,8 +1,6 @@
 package com.jof.springmvc.dao;
 
-import java.math.BigInteger;
-import java.util.List;
-
+import com.jof.springmvc.model.Review;
 import org.hibernate.Criteria;
 import org.hibernate.criterion.Order;
 import org.hibernate.criterion.Restrictions;
@@ -10,7 +8,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Repository;
 
-import com.jof.springmvc.model.Review;
+import java.util.List;
 
 
 @Repository("reviewDao")
@@ -36,7 +34,7 @@ public class ReviewDaoImpl extends AbstractDao<Integer, Review> implements Revie
     }
   
     @SuppressWarnings("unchecked")
-    public List<Review> findAllReviewByUser(BigInteger source_user_id) {
+    public List<Review> findAllReviewByUser(Long source_user_id) {
         Criteria criteria = createEntityCriteria().addOrder(Order.asc("id"));
         criteria.add(Restrictions.eq("source_user_id", source_user_id));
         criteria.setResultTransformer(Criteria.DISTINCT_ROOT_ENTITY);//To avoid duplicates.
@@ -47,7 +45,7 @@ public class ReviewDaoImpl extends AbstractDao<Integer, Review> implements Revie
     }
     
     @SuppressWarnings("unchecked")
-    public List<Review> findAllReviewAboutUser(BigInteger target_user_id) {
+    public List<Review> findAllReviewAboutUser(Long target_user_id) {
         Criteria criteria = createEntityCriteria().addOrder(Order.asc("id"));
         criteria.add(Restrictions.eq("target_user_id", target_user_id));
         criteria.setResultTransformer(Criteria.DISTINCT_ROOT_ENTITY);//To avoid duplicates.
