+++ src/main/java/com/jof/springmvc/service/ReviewService.java
@@ -1,9 +1,8 @@
 package com.jof.springmvc.service;
 
-import java.math.BigInteger;
-import java.util.List;
-
 import com.jof.springmvc.model.Review;
+
+import java.util.List;
 
 public interface ReviewService {
 
@@ -16,9 +15,9 @@ public interface ReviewService {
 
 	    void deleteById(int id);
 
-	    List<Review> findAllReviewByUser(BigInteger source_user_id);
+	    List<Review> findAllReviewByUser(Long source_user_id);
 	    
-	    List<Review> findAllReviewAboutUser(BigInteger target_user_id);
+	    List<Review> findAllReviewAboutUser(Long target_user_id);
 	    
 	    List<Review> findAllReviewForGame(int game_id);
